+++ index.php
@@ -55,4 +55,6 @@
 <!-- End article/class section -->
 
 
+
+
 <?php require_once('includes/footer.php'); ?>